feat: payment for membership subscription, continue modify the feature to check old member's subscription to discount

+++ src/main/java/com/smartquit/smartquitiot/dto/response/MembershipPackagePlan.java
@@ -12,7 +12,7 @@ import lombok.experimental.FieldDefaults;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class MembershipPackagePlan {
 
-    Double planPrice;
+    Long planPrice;
     Integer planDuration;
     String planDurationUnit;
     MembershipPackageDTO membershipPackage;
